Move HashType to /src/coin/baseCoin/iface.ts
HashType in src/utils/hash.ts cleaned and related unit tests refined
