+++ src/coin/xtz/utils.ts
@@ -4,7 +4,8 @@ import { InMemorySigner } from '@taquito/signer';
 import { ec as EC } from 'elliptic';
 import { SigningError } from '../baseCoin/errors';
 import * as Validator from '../../utils/validate';
-import { HashType, hashTypes } from '../../utils/hash';
+import { hashTypes } from '../../utils/hash';
+import { HashType } from '../baseCoin/iface';
 import { genericMultisigDataToSign } from './multisigUtils';
 import { SignResponse } from './iface';
 import { KeyPair } from './keyPair';
